et_viewer (et v1.7)

Displays a sequence chart for trace events (messages/actions)

Functions

file(FileName)

file(FileName) -> {ok, ViewerPid} | {error, Reason}

Start a new event viewer and a corresponding collector and load them with trace events from a trace file.

get_collector_pid(ViewerPid)

get_collector_pid(ViewerPid) -> CollectorPid

Returns the identifier of the collector process.

start()

start() -> ok

Simplified start of a sequence chart viewer with global tracing activated.

Convenient to be used from the command line (erl -s et_viewer).

start(GUI)

start(Options) -> ok

Start of a sequence chart viewer without linking to the parent process.

start_link(GUI)

start_link(Options) -> {ok, ViewerPid} | {error, Reason}

Start a sequence chart viewer for trace events (messages/actions)

A filter_fun() takes an event record as sole argument and returns false | true | {true, NewEvent}.

If the collector_pid is undefined a new et_collector will be started with the following parameter settings: parent_pid, event_order, trace_global, trace_pattern, trace_port, trace_max_queue, trace_client, dict_insert and dict_delete. The new et_viewer will register itself as an et_collector subscriber.

Default values:

parent_pid - self().

title - "et_viewer".

detail_level - max.

is_suspended - false.

scale - 2.

width - 800.

height - 600.

collector_pid - undefined.

event_order - trace_ts.

active_filter - collector.

max_actors - 5.

actors - ["UNKNOWN"].

first_event - first.

hide_unknown - false.

hide_actions - false.

display_mode - all.

stop(ViewerPid)

stop(ViewerPid) -> ok

Stops a viewer process.