et (et v1.6.5)

Interface module for the Event Trace (ET) application

Functions

phone_home(DetailLevel, FromTo, Label, Contents) -> hopefully_traced

These functions sends a signal to the outer space and the caller hopes that someone is listening. In other words, they invoke et:trace_me/4 and et:trace_me/5 respectively.

phone_home(DetailLevel, From, To, Label, Contents) -> hopefully_traced

report_event(DetailLevel, FromTo, Label, Contents) -> hopefully_traced

Deprecated functions which for the time being are kept for backwards compatibility. Invokes et:trace_me/4 and et:trace_me/5 respectively.

report_event(DetailLevel, From, To, Label, Contents) -> hopefully_traced

trace_me(DetailLevel, FromTo, Label, Contents) -> hopefully_traced
(since OTP R13B04)

Invokes et:trace_me/5 with both From and To set to FromTo.

trace_me(DetailLevel, From, To, Label, Contents) -> hopefully_traced
(since OTP R13B04)

DetailLevel = integer(X) when X =< 0, X >= 100

From = actor()

To = actor()

Label = atom() | string() | term()

Contents = [{Key, Value}] | term()

actor() = term()

A function that is intended to be traced.

This function is intended to be invoked at strategic places in user applications in order to enable simplified tracing. The functions are extremely light weight as they do nothing besides returning an atom. The functions are designed for being traced. The global tracing mechanism in et_collector defaults to set its trace pattern to these functions.

The label is intended to provide a brief summary of the event. It is preferred to use an atom but a string would also do.

The contents can be any term but in order to simplify post processing of the traced events, a plain list of {Key, Value} tuples is preferred.

Some events, such as messages, are directed from some actor to another. Other events (termed actions) may be undirected and only have one actor.